% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lamapoll_API_Wrapper.R
\name{backup.lamapoll}
\alias{backup.lamapoll}
\title{Lamapoll Data Download}
\usage{
backup.lamapoll(xstatus, xdir, ydir, apikey)
}
\arguments{
\item{xdir}{The directory into which the function will download the data, i.e. your backup folder.}

\item{ydir}{The directory that holds an excel file with one column, in which questionnaire names from Lamapoll are being pasted. You can use the function to download a selection of Lamapoll questionnaire data by specifying "selection" as status (xstatus) and creating an excel file with the dedicated questionnaire names in one column.}

\item{apikey}{The personal API key, that Lamapoll created for the DKJS User Account.}

\item{xtatus}{The status of the Questionnaire in Lamapoll. You can enter three types of status. 1., "selection". This will download the data of selected Questionnaires. The names of the questionnaires have to be entered in an excel file. Please ask Team W&E for further instructions. 2., "online". This will download all data of all questionnaires, that are currently online. 3., "offline". This will download all data of all questionnaires, that are currently offline. The package will create a subfolder for each type of status and save the data in the appropriate subfolder. The function will add one column to each datafile called "p_name", stating the name of the Lamapoll questionnaire.}
}
\value{
A folder that holds one excel file for each Lamapoll questionnaires with the specified status (xstatus).
}
\description{
This function Let's you download Data from Lamapoll through their API. Upon using, the package will create a folder in the path entered in the function (second parameter: "xdir"). The data will be stored in individual excel files for each questionnaire in this folder.
}
