% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lamapoll_API_Wrapper.R
\name{get.lamapoll.list}
\alias{get.lamapoll.list}
\title{Lamapoll Questionnaire List Download}
\usage{
get.lamapoll.list(xstatus, email, apikey)
}
\arguments{
\item{xstatus}{The status of the questionnaires of which you want all names. Can be set to "online" or "offline" or "inwork".}

\item{email}{The mailadress that you are registered with in Lamapoll.}

\item{apikey}{The personal API key, that Lamapoll created for you.}
}
\value{
An object of class dataframe.
}
\description{
This function Let's you download a list of questionnaire names from Lamapoll through their API. The result of the function is an object of class dataframe.
}
\examples{
downloaded_names <- get.lamapoll.list <-function("online", "duck@pond.com", "xyz1234")
}
