% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lamapoll_API_OF.R
\name{get_lamapoll_data}
\alias{get_lamapoll_data}
\title{Lamapoll Individual Data Download}
\usage{
get_lamapoll_data(email, poll_name, apikey)
}
\arguments{
\item{email}{The mailadress that you are registered with in Lamapoll.}

\item{poll_name}{The name of the questionnaire in Lamapoll, from which you want to download the data.}

\item{apikey}{The personal API key, that Lamapoll created for you.}
}
\value{
An object of class dataframe.
}
\description{
This function Let's you download Data of one questionnaire from Lamapoll through their API. The result of the function is an object of class dataframe.
}
\examples{
downloaded_data <- get_lamapoll_data <-function("duck@pond.com", "duckpoll", "xyz1234")
}
