% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lamapoll_API_OF.R
\name{get_lamapoll_list}
\alias{get_lamapoll_list}
\title{Lamapoll Questionnaire List Download}
\usage{
get_lamapoll_list(email, fb_status, apikey)
}
\arguments{
\item{email}{The mailadress that you are registered with in Lamapoll.}

\item{fb_status}{The status of the questionnaires of which you want all names. Can be set to "online" or "offline" or "inwork".}

\item{apikey}{The personal API key, that Lamapoll created for you.}
}
\value{
An object of class dataframe.
}
\description{
This function Let's you download a list of questionnaire names from Lamapoll through their API. The result of the function is an object of class dataframe.
}
\examples{
downloaded_names <- get_lamapoll_list <-function("duck@pond.com", "duckpoll", "xyz1234")
}
