% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lamapoll_API_Wrapper.R
\name{backup.lamapoll}
\alias{backup.lamapoll}
\title{Lamapoll Data Download}
\usage{
backup.lamapoll(xstatus, xdir, ydir, email, apikey)
}
\arguments{
\item{xdir}{The directory into which the function will download the data, i.e. your backup folder.}

\item{ydir}{The directory that holds an excel file with one column, in which questionnaire names from Lamapoll are being pasted. \cr
The directory has a maximum path length of ca. 80 characters due to MS Windows restrictions. With longer path names, the function will fail.
You can use the function to download a selection of Lamapoll questionnaire data by specifying "selection" as status (xstatus) and creating an excel file with the dedicated questionnaire names in one column. \cr
A directory has to be specified in the function, even when not utilizing the functions capability for the status "selection".}

\item{email}{The email address / username in Lamapoll for the DKJS User Account. Please make sure in Lamapoll, that the Account has the necessary rights.}

\item{apikey}{The personal API key, that Lamapoll created for the DKJS User Account.}

\item{xtatus}{The status of the Questionnaire in Lamapoll. You can enter three types of status.
\itemize {
    \item 1., "selection". This will download the data of selected Questionnaires. The names of the questionnaires have to be entered in an excel file, that is stored in the directory of (third parameter: "ydir").
    \item 2., "online". This will download all data of all questionnaires, that are currently online.
    \item 3., "offline". This will download all data of all questionnaires, that are currently offline.
    } \cr
 The function will create a subfolder for each type of status and save the data in a subfolder in the directory stated in xdir. \cr
 The function will also add one column to each datafile called "p_name" (i.e. "Programmname") , stating the name of the Lamapoll questionnaire.}
}
\value{
A folder that holds one excel file for each Lamapoll questionnaires with the specified status.
}
\description{
This function Let's you download Data from Lamapoll through their API. Upon using, the package will create a folder in the path entered in the function (second parameter: "xdir"). The data will be stored in individual excel files for each questionnaire in this folder.
}
\examples{
backup.lamapoll("online", "...Data/Backup/", "...Data/Programmliste.xslx", "duck@lamapoll.de", "xyz123") \cr
## Running this command will first internally generate a list of all questionnaires in Lamapoll, that are marked as online. \cr
## Then download all datafiles from these questionnaires into the Computers RAM. Then create a backupfolder to save these datafiles in. Then save these datafiles as individual .xlsx files.
}
