% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Lamapoll_API_Wrapper.R
\name{get.lamapoll.data}
\alias{get.lamapoll.data}
\title{Lamapoll Individual Data Download}
\usage{
get.lamapoll.data(poll_name, email, apikey)
}
\arguments{
\item{poll_name}{The name of the questionnaire in Lamapoll, from which you want to download the data.}

\item{email}{The mailadress that you are registered with in Lamapoll.}

\item{apikey}{The personal API key, that Lamapoll created for you.}
}
\value{
An object of class dataframe.
}
\description{
This function Let's you download Data of one questionnaire from Lamapoll through their API. The result of the function is an object of class dataframe.
}
\examples{
downloaded_data <- get.lamapoll.data <-function("duckpoll", "duck@pond.com", "xyz1234")
}
